refactor(exception): BaseException 위치 변경
