+++ server/src/main/java/com/onetool/server/api/order/service/OrderServiceImpl.java
@@ -4,21 +4,15 @@ import com.onetool.server.api.blueprint.Blueprint;
 import com.onetool.server.api.blueprint.repository.BlueprintRepository;
 import com.onetool.server.api.order.OrderBlueprint;
 import com.onetool.server.api.order.Orders;
-import com.onetool.server.api.order.dto.request.OrderItem;
 import com.onetool.server.api.order.dto.request.OrderRequest;
 import com.onetool.server.api.order.dto.response.OrderResponse;
 import com.onetool.server.api.order.repository.OrderRepository;
-import com.onetool.server.api.payments.dto.DepositResponse;
-import com.onetool.server.api.payments.repository.DepositRepository;
 import com.onetool.server.api.payments.service.DepositService;
-import com.onetool.server.global.auth.MemberAuthContext;
-import com.onetool.server.global.exception.BaseException;
+import com.onetool.server.global.exception.base.BaseException;
 import com.onetool.server.api.member.domain.Member;
 import com.onetool.server.api.member.repository.MemberRepository;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.hibernate.query.Order;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
