+++ server/src/main/java/com/onetool/server/api/payments/service/DepositService.java
@@ -3,11 +3,10 @@ package com.onetool.server.api.payments.service;
 import com.onetool.server.api.order.Orders;
 import com.onetool.server.api.order.repository.OrderRepository;
 import com.onetool.server.api.payments.domain.Payment;
-import com.onetool.server.api.payments.domain.PaymentBlueprint;
 import com.onetool.server.api.payments.dto.DepositRequest;
 import com.onetool.server.api.payments.dto.DepositResponse;
 import com.onetool.server.api.payments.repository.DepositRepository;
-import com.onetool.server.global.exception.BaseException;
+import com.onetool.server.global.exception.base.BaseException;
 import com.onetool.server.global.exception.codes.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
