+++ server/src/main/java/com/onetool/server/global/exception/MemberNotFoundException.java
@@ -1,5 +1,6 @@
 package com.onetool.server.global.exception;
 
+import com.onetool.server.global.exception.base.BaseException;
 import com.onetool.server.global.exception.codes.BaseCode;
 
 public class MemberNotFoundException extends BaseException {
