+++ server/src/main/java/com/onetool/server/global/exception/base/DuplicateException.java
@@ -1,0 +1,5 @@
+package com.onetool.server.global.exception.base;
+
+public class DuplicateException extends BaseException {
+
+}
